split out zip-table-of-contents-based hashing
