+++ src/rhash/rc_hash_internal.h
@@ -2,14 +2,38 @@
 #define RC_HASH_INTERNAL_H
 
 #include "rc_hash.h"
+#include "md5.h"
 
 RC_BEGIN_C_DECLS
+
+/* hash.c */
 
 void rc_hash_verbose(const rc_hash_callbacks_t* callbacks, const char* message);
 void rc_hash_verbose_formatted(const rc_hash_callbacks_t* callbacks, const char* format, ...);
 
 int rc_hash_error(const rc_hash_callbacks_t*, const char* message);
 int rc_hash_error_formatted(const rc_hash_callbacks_t* callbacks, const char* format, ...);
+
+void* rc_file_open(const rc_hash_iterator_t* iterator, const char* path);
+void rc_file_seek(const rc_hash_iterator_t* iterator, void* file_handle, int64_t offset, int origin);
+int64_t rc_file_tell(const rc_hash_iterator_t* iterator, void* file_handle);
+size_t rc_file_read(const rc_hash_iterator_t* iterator, void* file_handle, void* buffer, int requested_bytes);
+void rc_file_close(const rc_hash_iterator_t* iterator, void* file_handle);
+int64_t rc_file_size(const rc_hash_iterator_t* iterator, const char* path);
+
+
+void rc_hash_iterator_verbose(const rc_hash_iterator_t* iterator, const char* message);
+void rc_hash_iterator_verbose_formatted(const rc_hash_iterator_t* iterator, const char* format, ...);
+int rc_hash_iterator_error(const rc_hash_iterator_t* iterator, const char* message);
+int rc_hash_iterator_error_formatted(const rc_hash_iterator_t* iterator, const char* format, ...);
+
+
+int rc_hash_finalize(const rc_hash_iterator_t* iterator, md5_state_t* md5, char hash[33]);
+
+
+const char* rc_path_get_filename(const char* path);
+int rc_path_compare_extension(const char* path, const char* ext);
+
 
 typedef void (RC_CCONV* rc_hash_iterator_ext_handler_t)(rc_hash_iterator_t* iterator, const char* path, int data);
 typedef struct rc_hash_iterator_ext_handler_entry_t {
@@ -20,8 +44,6 @@ typedef struct rc_hash_iterator_ext_handler_entry_t {
 
 const rc_hash_iterator_ext_handler_entry_t* rc_hash_get_iterator_ext_handlers(size_t* num_handlers);
 
-const char* rc_path_get_filename(const char* path);
-int rc_path_compare_extension(const char* path, const char* ext);
 
 typedef struct rc_hash_cdrom_track_t {
   void* file_handle;        /* the file handle for reading the track data */
@@ -36,6 +58,9 @@ typedef struct rc_hash_cdrom_track_t {
   uint32_t track_id;        /* the index of the track */
 #endif
 } rc_hash_cdrom_track_t;
+
+/* hash_zip.c */
+int rc_hash_ms_dos(char hash[33], const rc_hash_iterator_t* iterator);
 
 RC_END_C_DECLS
