+++ test/rcheevos-test.vcxproj
@@ -159,6 +159,7 @@
     <ClCompile Include="..\src\rhash\aes.c" />
     <ClCompile Include="..\src\rhash\cdreader.c" />
     <ClCompile Include="..\src\rhash\hash.c" />
+    <ClCompile Include="..\src\rhash\hash_zip.c" />
     <ClCompile Include="..\src\rhash\md5.c" />
     <ClCompile Include="..\src\rurl\url.c" />
     <ClCompile Include="lua\src\lapi.c" />
@@ -217,6 +218,7 @@
     <ClCompile Include="rhash\mock_filereader.c" />
     <ClCompile Include="rhash\test_cdreader.c" />
     <ClCompile Include="rhash\test_hash.c" />
+    <ClCompile Include="rhash\test_hash_zip.c" />
     <ClCompile Include="rurl\test_url.c" />
     <ClCompile Include="test.c" />
     <ClCompile Include="test_rc_client.c" />
