+++ test/rhash/data.c
@@ -828,135 +828,3 @@ uint8_t* convert_to_2352(uint8_t* input, size_t* size, uint32_t first_sector)
     return output;
 }
 
-uint8_t* generate_zip_file(size_t* image_size)
-{
-  const uint8_t data_zip[] = {
-    'P','K',0x03,0x04, /* local file header signature */
-    0x14,0x00, 0x02,0x00, 0x08,0x00, 0x00,0xBC, 0x98,0x21, 0x00,0x00,0x00,0x00, 0x02,0x00,0x00,0x00, 0x00,0x00,0x00,0x00, /* info */
-    0x07,0x00, 0x00,0x00, 'F','O','L','D','E','R','/', /* file name length, extra length, name data */
-    0x03, 0x00, /* compressed content */
-
-    'P','K',0x03,0x04, /* local file header signature */
-    0x14,0x00, 0x02,0x00, 0x08,0x00, 0x00,0xBC, 0x98,0x21, 0x31,0xCF,0xD0,0x4A, 0x03,0x00,0x00,0x00, 0x01,0x00,0x00,0x00, /* info */
-    0x0E,0x00, 0x00,0x00, 'F','O','L','D','E','R','/','S','U','B','.','T','X','T', /* file name length, extra length, name data */
-    0x73, 0x02, 0x00, /* compressed content */
-
-    'P','K',0x03,0x04, /* local file header signature */
-    0x14,0x00, 0x02,0x00, 0x08,0x00, 0x00,0xBC, 0x98,0x21, 0x8B,0x9E,0xD9,0xD3, 0x03,0x00,0x00,0x00, 0x01,0x00,0x00,0x00, /* info */
-    0x08,0x00, 0x00,0x00, 'R','O','O','T','.','T','X','T', /* file name length, extra length, name data */
-    0x73, 0x04, 0x00, /* compressed content */
-
-    'P', 'K',0x01,0x02, /* central directory file header signature */
-    0x00,0x00, 0x14,0x00, 0x02,0x00, 0x08,0x00, 0x00,0xBC, 0x98,0x21, 0x00,0x00,0x00,0x00, 0x02,0x00,0x00,0x00, 0x00,0x00,0x00,0x00,
-    0x07,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00,0x00,0x00, 0x00,0x00,0x00,0x00, 
-    'F','O','L','D','E','R','/',
-
-    'P','K',0x01,0x02, /* central directory file header signature */
-    0x00,0x00, 0x14,0x00, 0x02,0x00, 0x08,0x00, 0x00,0xBC, 0x98,0x21, 0x31,0xCF,0xD0,0x4A, 0x03,0x00,0x00,0x00, 0x01,0x00,0x00,0x00,
-    0x0E,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00,0x00,0x00, 0x27,0x00,0x00,0x00,
-    'F','O','L','D','E','R','/','S','U','B','.','T','X','T',
-
-    'P','K',0x01,0x02, /* central directory file header signature */
-    0x00,0x00, 0x14,0x00, 0x02,0x00, 0x08,0x00, 0x00,0xBC, 0x98,0x21, 0x8B,0x9E,0xD9,0xD3, 0x03,0x00,0x00,0x00, 0x01,0x00,0x00,0x00,
-    0x08,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00,0x00,0x00, 0x56,0x00,0x00,0x00, 
-    'R','O','O','T','.','T','X','T',
-
-    'P','K',0x05,0x06, /* end of central directory signature */
-    0x00,0x00, 0x00,0x00, /* disk number */
-    0x03,0x00, 0x03,0x00, /* number of directory entries on this disk and total */
-    0xA7,0x00,0x00,0x00, /* size of central directory (bytes) */
-    0x7F,0x00,0x00,0x00, /* offset of start of central directory */
-    0x16,0x00, 'T','O','R','R','E','N','T','Z','I','P','P','E','D','-','F','D','0','7','C','5','2','C' /* comment length and comment data */
-  };
-
-  uint8_t* image = (uint8_t*)malloc(sizeof(data_zip));
-  if (image != NULL)
-    memcpy(image, data_zip, sizeof(data_zip));
-  if (image_size)
-    *image_size = sizeof(data_zip);
-  return image;
-}
-
-uint8_t* generate_zip64_file(size_t* image_size)
-{
-  const uint8_t data_zip64[] = {
-    'P','K',0x03,0x04, /* local file header signature */
-    0x14,0x00, 0x00,0x00, 0x08,0x00, 0x30,0x74, 0x0A,0x41, 0x7E,0xE7,0xFF,0x69, 0x24,0x00,0x00,0x00, 0x24,0x00,0x00,0x00,
-    0x06,0x00, 0x00,0x00, 'R','E','A','D','M','E',
-    0x0B,0xC9,0xC8,0x2C,0x56,0x28,0xCE,0x4D,0xCC,0xC9,0x51,0x48,0xCB,0xCC,0x49,0x55,0x00,0xF2,
-    0x32,0xF3,0x14,0xA2,0x3C,0x03,0xCC,0x4C,0x14,0xD2,0xF2,0x8B,0x72,0x13,0x4B,0xF4,0xB8,0x00,
-
-    'P','K',0x01,0x02, /* central directory file header signature */
-    0x2D,0x03, 0x2D,0x00, 0x00,0x00, 0x08,0x00, 0x30,0x74, 0x0A,0x41, 0x7E,0xE7,0xFF,0x69, 0xFF,0xFF,0xFF,0xFF, 0xFF,0xFF,0xFF,0xFF,
-    0x06,0x00, 0x14,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00,0xA4,0x81, 0x00,0x00,0x00,0x00,
-    'R','E','A','D','M','E',
-    0x01,0x00, /* Zip64 extended information extra field header id */
-    0x10,0x00, /* Size of extra field chunk */
-    0x24,0x00,0x00,0x00,0x00,0x00,0x00,0x00, /* uncompressed file size */
-    0x24,0x00,0x00,0x00,0x00,0x00,0x00,0x00, /* size of compressed data */
-
-    'P','K',0x06,0x06, /* Zip64 End of central directory record signature */
-    0x2C,0x00,0x00,0x00,0x00,0x00,0x00,0x00, /* Size of the EOCD64 minus 12 */
-    0x2D,0x00, 0x2D,0x00, 0x00,0x00,0x00,0x00, 0x00,0x00,0x00,0x00,
-    0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x48,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x48,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-
-    'P','K',0x06,0x07, /* end of central dir locator signature */
-    0x00,0x00,0x00,0x00, /* number of the disk with the start of the zip64 end of central directory */
-    0x90,0x00,0x00,0x00,0x00,0x00,0x00,0x00, /* relative offset of the zip64 end of central directory record */
-    0x01,0x00,0x00,0x00, /* total number of disks */
-
-    'P','K',0x05,0x06, /* end of central directory signature */
-    0x00,0x00, 0x00,0x00, /* disk number */
-    0xFF,0xFF,0xFF,0xFF,0xFF,0xFF,0xFF,0xFF,0xFF,0xFF,0xFF,0xFF, /* file and size numbers are 0xFF in Zip64 */
-    0x00,0x00 /* comment length */
-  };
-
-  uint8_t* image = (uint8_t*)malloc(sizeof(data_zip64));
-  if (image != NULL)
-    memcpy(image, data_zip64, sizeof(data_zip64));
-  if (image_size)
-    *image_size = sizeof(data_zip64);
-  return image;
-}
-
-uint8_t* generate_child_dosz_file(size_t* image_size)
-{
-  const uint8_t data_zip[] = {
-    'P','K',0x03,0x04, /* local file header signature */
-    0x0A,0x00, 0x00,0x00, 0x00,0x00, 0x00,0xBC, 0x98,0x21, 0x00,0x00,0x00,0x00, 0x00,0x00,0x00,0x00, 0x00,0x00,0x00,0x00, /* info */
-    0x10,0x00, 0x00,0x00, 'b','a','s','e','.','d','o','s','z','.','p','a','r','e','n','t', /* file name length; extra length; name data */
-    0x73, 0x04, 0x00, /* compressed content */
-
-    'P','K',0x03,0x04, 
-    0x0A,0x00, 0x00,0x00, 0x00,0x00, 0x00,0xBC, 0x98,0x21, 0x29,0x54,0xB3,0x22, 0x05,0x00,0x00,0x00, 0x05,0x00,0x00,0x00, /* info */
-    0x09,0x00, 0x00,0x00, 'C','H','I','L','D','.','T','X','T', /* file name length; extra length; name data */
-    0x63, 0x68, 0x69, 0x6C, 0x64, /* content */
-
-    'P','K',0x01,0x02, /* central directory file header signature */
-    0x3F,0x00, 0x0A,0x00, 0x00,0x00, 0x00,0x00, 0x00,0xBC, 0x98,0x21, 0x00,0x00,0x00,0x00, 0x00,0x00,0x00,0x00, 0x00,0x00,0x00,0x00,
-    0x10,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x20,0x00,0x00,0x00, 0x00,0x00,0x00,0x00,
-    'b','a','s','e','.','d','o','s','z','.','p','a','r','e','n','t',
-
-    'P','K',0x01,0x02, /* central directory file header signature */
-    0x3F,0x00, 0x0A,0x00, 0x00,0x00, 0x00,0x00, 0x00,0xBC, 0x98,0x21, 0x29,0x54,0xB3,0x22, 0x05,0x00,0x00,0x00, 0x05,0x00,0x00,0x00,
-    0x09,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x00,0x00, 0x20,0x00,0x00,0x00, 0x00,0x00,0x00,0x00,
-    'C','H','I','L','D','.','T','X','T',
-
-    'P','K',0x05,0x06, /* end of central directory signature */
-    0x00,0x00, 0x00,0x00, /* disk number */
-    0x02,0x00, 0x02,0x00, /* number of directory entries on this disk and total */
-    0x75,0x00,0x00,0x00, /* size of central directory (bytes) */
-    0x5D,0x00,0x00,0x00, /* offset of start of central directory */
-    0x00,0x00 /* comment length */
-  };
-
-  uint8_t* image = (uint8_t*)malloc(sizeof(data_zip));
-  if (image != NULL)
-    memcpy(image, data_zip, sizeof(data_zip));
-  if (image_size)
-    *image_size = sizeof(data_zip);
-  return image;
-}
