+++ test/test.c
@@ -82,6 +82,7 @@ extern void test_url();
 
 extern void test_cdreader();
 extern void test_hash();
+extern void test_hash_zip();
 
 extern void test_rapi_common();
 extern void test_rapi_user();
@@ -134,6 +135,7 @@ int main(void) {
   test_rc_libretro(); /* libretro extensions require hash support */
   test_cdreader();
   test_hash();
+  test_hash_zip();
 #endif
 #endif
